requirements.txt
# FastAPI и сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==4.6.0
hiredis==2.2.3

# Celery
celery==5.3.4
# celery[redis] не нужен отдельно, так как redis уже указан выше

# Аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Валидация и схемы
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Утилиты
python-dotenv==1.0.0
aiofiles==23.2.1

# PDF генерация
weasyprint==60.2
reportlab==4.0.7

# Оптимизация изображений
Pillow==10.1.0

# Логирование
structlog==23.2.0

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
aiosqlite==0.19.0
pytest-benchmark==4.0.0

# Документация
# Автоматически через FastAPI

